[Refactor] search textField & topBar
