+++ app/src/main/java/com/awilab/movieencyclopedia/ui/movie/MovieScreen.kt
@@ -4,14 +4,13 @@ import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
+import com.awilab.movieencyclopedia.R
+import com.awilab.movieencyclopedia.ui.widgets.Appbar
 import com.ramcosta.composedestinations.annotation.Destination
 import com.ramcosta.composedestinations.annotation.RootGraph
 import org.koin.androidx.compose.koinViewModel
@@ -25,12 +24,7 @@ fun MovieScreen(
 ) {
     Scaffold(
         topBar = {
-            TopAppBar(
-                colors = TopAppBarDefaults.centerAlignedTopAppBarColors(
-                    containerColor = MaterialTheme.colorScheme.primary,
-                ),
-                title = { Text("MovieScreen", color = MaterialTheme.colorScheme.onPrimaryContainer) },
-            )
+            Appbar(titleRes = R.string.lab_movie)
         },
     ) { innerPadding ->
         Column(
