+++ app/src/main/java/com/awilab/movieencyclopedia/ui/search/SearchScreen.kt
@@ -1,6 +1,5 @@
 package com.awilab.movieencyclopedia.ui.search
 
-import androidx.annotation.StringRes
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
@@ -24,6 +23,8 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
@@ -48,13 +49,12 @@ fun SearchScreen(
                 modifier = Modifier
                     .padding(innerPadding)
                     .fillMaxWidth()
-                    .background(color = MaterialTheme.colorScheme.primary),
+                    .background(color = MaterialTheme.colorScheme.secondaryContainer),
             ) {
                 SearchField(
                     keyword = vm.keywordStateFlow.collectAsState(initial = "").value,
                     onValueChange = vm::onSearch,
                     onClear = vm::onClear,
-                    errorMsg = vm.onWarningMsg(),
                 )
             }
 
@@ -68,11 +68,10 @@ fun SearchField(
     keyword: String,
     onValueChange: (keyword: String) -> Unit,
     onClear: () -> Unit,
-    @StringRes errorMsg: Int,
 ) {
     OutlinedTextField(
         modifier = Modifier
-            .padding(horizontal = 16.dp, vertical = 8.dp)
+            .padding(start = 16.dp, end = 16.dp, bottom = 8.dp)
             .fillMaxWidth(),
         shape = MaterialTheme.shapes.extraLarge,
         value = keyword,
@@ -103,21 +102,12 @@ fun SearchField(
                 }
             }
         },
-        supportingText = {
-            AnimatedVisibility(
-                visible = keyword.isNotBlank(),
-                enter = fadeIn() + scaleIn(),
-                exit = fadeOut() + scaleOut(),
-            ) {
-                Text(text = stringResource(id = errorMsg), color = MaterialTheme.colorScheme.error)
-            }
-        },
         colors = OutlinedTextFieldDefaults.colors(
-            focusedBorderColor = MaterialTheme.colorScheme.outlineVariant,
+            focusedBorderColor = MaterialTheme.colorScheme.onSecondaryContainer,
             unfocusedBorderColor = MaterialTheme.colorScheme.outline,
-            focusedLabelColor = MaterialTheme.colorScheme.outlineVariant,
+            focusedLabelColor = MaterialTheme.colorScheme.onSecondaryContainer,
             unfocusedLabelColor = MaterialTheme.colorScheme.outline,
-            cursorColor = MaterialTheme.colorScheme.outline,
+            cursorColor = MaterialTheme.colorScheme.onSecondaryContainer,
         ),
     )
 }
