+++ app/src/main/java/com/awilab/movieencyclopedia/ui/widgets/AppBar.kt
@@ -2,28 +2,39 @@
 
 package com.awilab.movieencyclopedia.ui.widgets
 
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.KeyboardArrowLeft
-import androidx.compose.material.icons.filled.KeyboardArrowLeft
-import androidx.compose.material3.CenterAlignedTopAppBar
+import android.annotation.SuppressLint
+import androidx.annotation.StringRes
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
 
-fun appBar(title: String): @Composable () -> Unit = { TopAppBar(title = { Text(text = title) }) }
-
+@StringRes
 @Composable
-fun AppBarWithBackBtn(title: String, backEvent: () -> Unit) {
-    CenterAlignedTopAppBar(
-        title = { Text(text = title) },
-        navigationIcon = {
-            IconButton(onClick = { /*TODO*/ }) {
-                Icon(
-                    imageVector = Icons.AutoMirrored.Filled.KeyboardArrowLeft,
-                    contentDescription = "",
+@SuppressLint("SupportAnnotationUsage")
+fun Appbar(titleRes: Int) {
+    TopAppBar(
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(72.dp),
+        colors = TopAppBarDefaults.centerAlignedTopAppBarColors(
+            containerColor = MaterialTheme.colorScheme.secondaryContainer,
+        ),
+        title = {
+            Column(modifier = Modifier.fillMaxSize(), verticalArrangement = Arrangement.Center) {
+                Text(
+                    text = stringResource(id = titleRes),
+                    color = MaterialTheme.colorScheme.onSecondaryContainer,
                 )
             }
         },
